move flight state to performace model
