--- 1682/gas_male/submodels/flight_state.py
@@ -1,46 +1,0 @@
-" flight_state.py "
-from gpkit import Model, Variable
-
-class FlightState(Model):
-    "define environment state during a portion of an aircraft mission"
-    def __init__(self, alt, onStation, wind, **kwargs):
-
-        self.onStation = onStation
-
-        mu = Variable("\\mu", 1.628e-5, "N*s/m^2", "dynamic viscosity")
-        rho = Variable("\\rho", "kg/m^3", "air density")
-        h = Variable("h", alt, "ft", "altitude")
-        href = Variable("h_{ref}", 15000, "ft", "Reference altitude")
-        psl = Variable("p_{sl}", 101325, "Pa", "Pressure at sea level")
-        Latm = Variable("L_{atm}", 0.0065, "K/m", "Temperature lapse rate")
-        Tsl = Variable("T_{sl}", 288.15, "K", "Temperature at sea level")
-        temp = [(t.value - l.value*v.value).magnitude
-                for t, v, l in zip(Tsl, h, Latm)]
-        Tatm = Variable("t_{atm}", temp, "K", "Air temperature")
-        mu = Variable("\\mu", "N*s/m^2", "Dynamic viscosity")
-        musl = Variable("\\mu_{sl}", 1.789*10**-5, "N*s/m^2",
-                        "Dynamic viscosity at sea level")
-        Rspec = Variable("R_{spec}", 287.058, "J/kg/K",
-                         "Specific gas constant of air")
-
-        # Atmospheric variation with altitude (valid from 0-7km of altitude)
-        constraints = [rho == psl*Tatm**(5.257-1)/Rspec/(Tsl**5.257),
-                       (mu/musl)**0.1 == 0.991*(h/href)**(-0.00529),
-                       h == h,
-                       href == href]
-
-        V = Variable("V", "m/s", "true airspeed")
-
-        if wind:
-
-            V_wind = Variable("V_{wind}", 25, "m/s", "Wind speed")
-            constraints.extend([V >= V_wind])
-
-        else:
-
-            V_wind = Variable("V_{wind}", "m/s", "Wind speed")
-            V_ref = Variable("V_{ref}", 25, "m/s", "Reference wind speed")
-
-            constraints.extend([(V_wind/V_ref) >= 0.6462*(h/href) + 0.3538,
-                                V >= V_wind])
-        Model.__init__(self, None, constraints, **kwargs)
